Bugfixes regarding issue  #8
and some minor changes

+++ Arduino/McLighting/mode_custom_ws2812fx_animations.h
@@ -58,25 +58,25 @@ void Fire2012() {
 // Array of temperature readings at each simulation cell
   
   // Step 1.  Cool down every cell a little
-  for( int i = 0; i < WS2812FXStripSettings.stripSize; i++) {
+  for( uint16_t i = 0; i < WS2812FXStripSettings.stripSize; i++) {
     ledstates[i] = qsub8( ledstates[i],  random8(0, ((COOLING * 10) / WS2812FXStripSettings.stripSize) + 2));
   }
 
   // Step 2.  Heat from each cell drifts 'up' and diffuses a little
-  for( int k= WS2812FXStripSettings.stripSize - 1; k >= 2; k--) {
+  for( uint16_t k= WS2812FXStripSettings.stripSize - 1; k >= 2; k--) {
     ledstates[k] = (ledstates[k - 1] + ledstates[k - 2] + ledstates[k - 2]) / 3;
   }
 
   // Step 3.  Randomly ignite new 'sparks' of heat near the bottom
   if( random8() < SPARKING ) {
-    int y = random8(7);
+    uint8_t y = random8(7);
     ledstates[y] = qadd8(ledstates[y], random8(160,255) );
   }
 
   // Step 4.  Map from heat cells to LED colors
-  for( int j = 0; j < WS2812FXStripSettings.stripSize; j++) {
+  for( uint16_t j = 0; j < WS2812FXStripSettings.stripSize; j++) {
     CRGB color = HeatColor( ledstates[j]);
-    int pixelnumber;
+    uint16_t pixelnumber;
     if( gReverseDirection ) {
       pixelnumber = (WS2812FXStripSettings.stripSize - 1) - j;
     } else {
